+++ src/main/java/net/collaud/fablab/manager/data/ConfigurationEO.java
@@ -28,13 +28,13 @@ import net.collaud.fablab.manager.data.type.ConfigurationKey;
 @Slf4j
 @AllArgsConstructor
 @NoArgsConstructor
-public class ConfigurationEO extends AbstractDataEO<Integer> implements Serializable {
+public class ConfigurationEO extends AbstractDataEO<Long> implements Serializable {
 
 	@JsonIgnore
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id", nullable = false)
-	private Integer id;
+	private Long id;
 
 	@Column(name = "conf_key", nullable = false)
 	private String keyString;
