+++ src/main/java/net/collaud/fablab/manager/rest/v1/data/UserSimpleTO.java
@@ -3,7 +3,6 @@ package net.collaud.fablab.manager.rest.v1.data;
 import lombok.Getter;
 import lombok.Setter;
 import net.collaud.fablab.manager.data.MembershipTypeEO;
-import net.collaud.fablab.manager.data.SubscriptionEO;
 import net.collaud.fablab.manager.data.UserEO;
 
 /**
@@ -14,7 +13,7 @@ import net.collaud.fablab.manager.data.UserEO;
 @Setter
 public class UserSimpleTO extends AbstractTO<UserEO, UserSimpleTO> {
 
-	private Integer id;
+	private Long id;
 	private String firstname;
 	private String lastname;
 	private String email;
